requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.4
python-decouple==3.6
sqlparse==0.4.3
tzdata==2022.4
